rename: tsx파일 파스칼 통일, 이미지 파일 분류
