+++ frontend/public/images/product/icon-delete.png
