+++ frontend/src/components/product/Photos.tsx
