+++ frontend/src/routes/pages/Product.tsx
@@ -1,8 +1,8 @@
-import Photos from "@/components/product/PhotoSection";
-import DetailsSection from "@/components/product/detailsSection";
-import DetailsTab from "@/components/product/detailsTabSection";
+import Photos from "@/components/product/Photos";
+import DetailsSection from "@/components/product/Details";
+import DetailsTab from "@/components/product/DetailsTab";
 import ReviewsTab from "@/components/product/reviewsTab";
-import QnasTab from "@/components/product/faqsTab";
+import QnasTab from "@/components/product/QnasTab";
 import ExchangeTab from "@/components/product/exchangeTab";
 
 function Product() {
